+++ packages/core/package.json
@@ -10,8 +10,6 @@
   },
   "dependencies": {
     "body-parser": "^1.20.3",
-    "esbuild": "^0.24.2",
-    "esbuild-plugin-tsc": "^0.4.0",
     "express": "^4.21.2",
     "ioredis": "^5.4.2",
     "pino": "^9.6.0",
